update: refactor and separate kenkou_setting.go
