+++ internal/db/connect.go
@@ -3,6 +3,8 @@ package db
 import (
 	"database/sql"
 	"log"
+
+	_ "github.com/mattn/go-sqlite3"
 )
 
 func Connect() *sql.DB {
